[refactor] wiki ツリーの DFS 整形を flatten_wiki_tree に集約
tui._flatten_wiki_tree と api.list_wikis の print_tree で同じツリー
装飾ロジックを重複実装していたので、api 側に flatten_wiki_tree を
切り出して双方から使う形に整理。戻り値は (page 辞書, ツリー前置子)
のペア列で、呼び出し側がタイトルや URL を後付けする。
